Upgraded protobuf to support cross crate imports. (#24)
Upgraded protobuf to support cross crate imports.
We need it to allow defining protos in zksync-era as well.
Proto packages share a single common namespace, so I prefixed our
packages with zksync.<crate name>.
I couldn't name the crate "protobuf", because cargo deny has mistaken it
for the public crate with the same name. I've named it
"zksync_protobuf", to be consistent with zksync-era repo.

CI protobuf compatibility check is now more complex, because we need to
collect the descriptors from the cargo build artifacts. Currently it is
implemented as a bash command, but can be upgraded to a proper github
action in the future.

---------

Co-authored-by: Bruno França <bruno@franca.xyz>