fix: Fix extra `GetBlock` requests for validators (#27)
# What ❔

If the `SyncBlocks` actor runs on a validator node, the node previously
produced unnecessary `GetBlock` requests to peers. This is because
`SyncBlocks` logic assumed that new blocks are only added within the
actor. This PR lifts this assumption; it filters `GetBlock` requests and
cancels pending requests reacting to new blocks being put in the
storage.

## Why ❔


Extra `GetBlock` requests waste traffic and processing time of peers.