Removed protocol_version from config (#49)
## What ❔

Removed protocol_version from config.
Instead, I've added a constant in bft crate indicating the version of
the protocol implementation.
The payload format should be versioned separately. Payload format
version should be included in the payload itself (set via
PayloadSource::propose, verified via WriteBlockStore::verify_payload).

## Why ❔

What is being versioned is actually the implementation of the consensus,
so it doesn't make sense to configure it via config. If it was
configurable the implementation wouldn't be able to reason about the
protocol version and compatibility of the peers.