Improvements on standalone node execution (#55)
## What ❔

Documented a guide on setting up and running a local consensus node,
along with a new Makefile containing commands for easy execution. Also
modifies the database directory path for each node.

## Why ❔

The README and the Makefile would streamline generating the
configuration for all nodes with a single command and simplify running
the desired node. About the modification for the database directory
path, it seems beneficial for running multiple nodes on a single
machine, especially for local testing. However, in the broader
architecture of the Test Framework, each node is expected to operate
within a distinct Docker container or a similar setup. In that context,
this change might not be useful.