Consensus node Dockerfile for test framework  (#57)
## What ❔

Add Dockerfile to create a container with a consensus node running in
it.

## Why ❔

This introduces a Dockerfile for the test framework proposal, enabling
the execution of a node in isolation for potential use in future
consensus among N nodes in the network. Additionally, a compose file is
included for testing purposes, facilitating the simultaneous launch of
two nodes that can initiate consensus, each running in a separate
container. This file might be replaced in the future with a script that
initializes N nodes, allowing for easier management, potentially with
Kubernetes or a similar tool.

---------

Co-authored-by: Grzegorz Prusak <pompon.pompon@gmail.com>