support for running a node from a state snapshot (BFT-418) (#75)
Support for a situation in which a node is started with a blockchain
state at block X > 0. Such a node doesn't want to fetch and store blocks
<X. I've adjusted syncing logic and validator logic (since validator
also might start from state snapshot).
Added tests for the new cases.

I've removed `parent` from the block struct, because the cost of
maintaining it exceeds the gains at this point (in zksync-era payload
implementation we have a commitment to the chain history anyway).