Clean old files for testing framework (#73)
## What ❔

Removing old files used to test the standalone nodes running locally and
updating the makefile targets to function in both the local and
Kubernetes environments.

## Why ❔

To maintain the node as standalone and ensure that the testing framework
remains organized and aligned with the latest changes. The node will no
longer run only in a dockerized environment and will always run inside a
pod in kubernetes so it's not necessary to keep all the files and
makefile targets to run the node only in Docker.