Node state for the RPC server (#67)
## What ❔

Make the `ReplicaState` struct and additional information about the node
state available on the RPC server side.

## Why ❔

For future tests, it would be necessary to have access to the state of
different running nodes in order to assess the behavior of the entire
consensus network. This could involve getting different metrics such as
the last saved view for each node, proposals, highest vote or quorum
certificate (QC), etc.