refactor(crypto): change hash_to_point algorithm (BFT-404) (#80)
## What ❔

Changing the existing `hash_to_point` algorithm to the standard
try-and-increment algorithm:
* hash msg into 32 uniform random bytes
* reduce the 32 bytes to base field element for the `x` coordinate
* try to get the `y` coordinate by computing the square root of `x^3 +
b`
* if failed, iterate again with `x + 1`

## Why ❔

The existing hash to curve algorithm, which turns 32 bytes to
`G1::compressed` and maps it into `G1::uncompressed` isn’t valid; it
doesn’t produce a point with an unknown discrete log.

## Notes
* The function changed to return the successful nonce, in addition to
the point. It’s currently unused, but suppose to be used for onchain
verification, so that it won't require a loop. From the benchmarks I've
run, the avg number of iteration is around 2. (nonce=1)
* Need to review the crypto security of this approach, along with the
expected onchain gas usage, mainly for the square root. I already
prepared a compatible Solidity implementation, but I haven’t checked gas
yet.
* Need to review whether 32 uniform random bytes, to be reduced to `x`,
are sufficient, or that it should be extended to 64 bytes, given the
onchain verification gas usage implications.
* Need to review whether just picking [only] the positive square root is
good enough.
* The prime field modulus
(`0x30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd47` /
`21888242871839275222246405745257275088696311157297823662689037894645226208583`)
is defined in the base field type macro definition, but I found no way
to access it, hence it is currently redefined in the implementation.
This should be fixed.
* `get_point_from_x` is already implemented for `G1Affine`, but it's
private and unaccessible. This should be fixed.

---------

Co-authored-by: Bruno França <bruno@franca.xyz>