Bumped protocol version (#89)
Recent change of hash_to_point implementation is not backward
compatible. I've bumped the protocol version and added tests so that in
the future we won't introduce backward incompatible protocol changes to
the crypto logic by accident. We are about to introduce several more
backward incompatible changes, and they will also be included into the
protocol version 1.