Weighted validators (BFT-408, BFT-424) (#77)
## What ❔

Add weights to the validator set and change the consensus threshold
rules to require >80% of weight.

## Why ❔

We need for the L1 contracts to only accept blocks that were built by
our node (which is the single leader). We can give more weight to our
node, so that a QC isn't valid unless it signs. In turn this will be
useful in the future