Return Result<(), Error> on QC add  (#112)
## What ❔

Refactored the code to return errors on `PrepareQC::add` and
`CommitQC::add`. We were just ignoring messages from different views, or
signed by replicas not in the committee, or signed twice. Now these
scenarios return an error.

## Why ❔

From a code review it was suggested to return errors in `PrepareQC::add`
and `CommitQC::add` instead of ignoring the message. FWIW, these cases
are already been checked by our current caller, but for correctness we
believe these functions should provide proper errors in case addition to
the QC is not performed.