BFT-463: Add LeaderSelectionMode::Rota (#115)
## What ❔

Adds a `LeaderSelectionMode::Rota` to allow specifying leaders in
subsequent views using a list of public keys. The 1st validator on the
list is going to be the leader in view 0, the 2nd in view 1, the 3rd in
view 2, wrapping around to start from the beginning as the views
increase. There can be repeated keys on the list, and they all have to
be part of the committee.

## Why ❔

This is part of BFT-439 to modify the BFT tests to use the approach
outlined in the [Twins paper](https://arxiv.org/abs/2004.10617). Part of
that is the ability to control who is the leader in each of the views.

With this change we should be able to generate a fixed leader schedule
for the first few rounds of a protocol. The implementation wraps around
so it's a total function, but in practice the tests will provide leaders
for all the rounds executed in the test.