made tests more deterministic (#129)
Some of the tests were flaky because of cpu contention affecting
interleavings. This fix makes the passing time (almost) deterministic,
by the use of `tokio::time::pause()`. The other consequence is that
runtime used in multiple tests has been changed from multi-threaded to
single-threaded. This also helps the determinism because the nextest
parallelism is bounded by the number of cores - if we make all tests
single-threaded, behavior of the test will be the same when executed
separately and in a test suite.

Also converted a should_panic test to regular test for consistency.
Tuned size of some of the tests to reasonable execution times.