feat: Batch vote publish loop (BFT-474) (#148)
## What ❔

- [x] Add `BatchStore::unsigned_batch_numbers` to query `BatchNumber`s
that still need a QC
- [x] Add `BatchStore::batch_to_sign` to fetch the `Batch` with the
correct `BatchHash` that needs to be signed
- [x] Add `PersistentBatchStore::unsigned_batch_numbers` to implement
the above
- [x] Add `PersistentBatchStore::batch_to_sign` to implement the above
- [x] Add `AttesterRunner` to periodically query
`unsigned_batch_numbers`, retrieve the data with `batch_to_sign` and
publish the `Signed<Batch>` to the gossip network.
- [x] Run the `AttesterRunner` in `Executor` if the node has an
`Attester` key.
- [x] Add a minimum batch number to sign to prevent repeated signing

The code will re-submit the node's signature after a restart to a number
of batches to be determined by the database query. We can say we'll
re-publish the top 10 unsigned batches, or we can say we'll only look
for the ones which follow the last finalised batch observed the Main
node. It is enough to publish a vote once because of the way
`BatchVotes` works (gossip peers will pull the data they need when they
connect). Because we'll always have the batches in order, we can take
note of the highest number we have published a vote for and never
publish it again (until a restart), to save ourselves the trouble of
repeatedly retrieving and signing the payload. That said it is assumed
that `batch_to_sign` is cheap as it just needs to look up a hash in the
DB.

## Why ❔

To publish the attestations over L1 batches to the gossip network, so
that a QC can be formed in
https://github.com/matter-labs/era-consensus/pull/145

---------

Co-authored-by: Bruno França <bruno@franca.xyz>