fix: Add max_batch_size to limit the SyncBatch request sizes (#152)
## What ❔

Adds a `max_batch_size` setting and uses it to set a `max_req_size`
limit in the RPC that deals with `SyncBatch`.

Fixes some of the comments added in
https://github.com/matter-labs/era-consensus/pull/150

### Alternative configuration

The size of `SyncBatch` depends on how many blocks are in the batch,
plus a fixed size proof. Since we already have a limit for the size of
block payloads, ostensibly we could add a limit on the number of blocks
in the batch; I used 100 in the protobuf reader if the field was
missing. Perhaps this setting is already available somewhere in
zksync-era?


## Why ❔

Tests failed in https://github.com/matter-labs/zksync-era/pull/2340 when
I unintentionally enabled the gossiping of `SyncBatch` due to the large
size of the requests. It turns out they were either hardcoded constants
copied from the block store state gossip (which doesn't depend on the
payload size, only the QC), or they wrongly used the block payload
limit.