refactor: Use earliest batch number to sign (BFT-474) (#150)
## What ❔

Followup for some design discussions in
https://github.com/matter-labs/era-consensus/pull/148

* Removes `BatchStore::unsigned_batch_numbers`
* Add `BatchStore::earliest_batch_number_to_sign`
* Add `BatchStore::latest_batch_number`
* Changes `AttesterRunner` to have a mutable range of
`earliest_batch_number` and `latest_batch_number` instead of querying
`unsigned_batch_numbers` and filtering it with a `min_batch_number`

## Why ❔

This might be more intuitive than the original version. 

Another motivation was that the current implementation of
`unsigned_batch_numbers` in
https://github.com/matter-labs/zksync-era/pull/2399 has to do two
queries: 1) to figure out the latest batch number and 2) to find all
batches without certificates within a limited range of the latest batch.

Here we only query the latest batch number in each loop, and we forego
the check on whether a QC already exists, otherwise we'd be back at
doing two queries, or some ugly book keeping to only check it on the
first loop after a restart. It should be cheap enough to publish a
duplicate vote.