feat: Batch vote metrics (BFT-486) (#157)
## What ❔

Adds metrics around the gossiping of batch votes and the storage of
batches.

Added `gossip::metrics::BATCH_VOTES_METRICS`:
- [x] `committee_size` to show how many attesters we have, which can be
correlated with the number of votes we receive over time. Wanted to add
a weight based one as well, but not sure if counting the weights
monotonically would not overflow the counter.
- [x] `votes_added` counts the votes received since start; with its
rate-of-change we can get an idea of how many attesters are producing
votes
- [x] `weight_added` counts the total weight of votes added since start,
each normalized to [0; 1] range with the total committee weight; this
way it shouldn't overflow, and its rate of growth should be 0.8-1.0 per
batch.
- [x] `min_batch_number` shows the minimum batch we expect votes for;
going up will indicate that QCs are formed
- [x] `last_added_vote_batch_number` is the number from the last vote we
registered; apart from Byzantine votes it should average to go up as L1
batches are created and votes are received, even if a QC cannot be
formed
- [x] `last_signed_batch_number` indicates that the current node is
publishing its votes and keeping up with the L1 batches

Also implement metrics for the `BatchStore` similar to how the
`BlockStore` has them:
- [x] `next` for `queued` and `persisted`
- [x] latencies for the `PersistedBatchStore` operations
- [x] latencies for the `wait_for_...` methods
- [x] last batch QC saved



## Why ❔

To make the voting process observable.