improve: AttestationStatusWatch::next_batch_to_attest non-optional (BFT-496) (#165)
## What ❔

`AttestationStatusWatch` must be initialised with a `BatchNumber`, it
cannot have `None` any more. `AttestationStatusRunner::new` was replaced
with the `AttestationStatusRunner::init` method which asynchronously
polls the API until the first value is returned, and then returns itself
along with the `AttestationStatusWatch` it created. This can then be
passed to the `Executor`, while the `AttestationStatusRunner::run` will
keep the status up to date in the background.

## Why ❔

In the review of https://github.com/matter-labs/era-consensus/pull/161
it was observed that the `Executor` can wait until this data is
available. In theory it is only unavailable if the main node API is
down, in which case an external node couldn't pull Genesis either and
would probably fail during startup, or if the Genesis itself is still
under construction in the database, which is a transient state under
which an external node as mentioned would not start, and apparently the
main node doesn't need the `Executor` to get over it. By removing `None`
as an option for `next_batch_to_attest` the state is easier to reason
about.