fix: AttestationStatusRunner::poll_until_some return after success (BFT-496) (#166)
## What ❔

Fixes `AttestationStatusRunner::poll_until_some` to return after a
successful poll. Originally it returned a value, then decided to write
it into the field and forgot to add a return statement.

## Why ❔

Tests in https://github.com/matter-labs/zksync-era/pull/2544 time out
because the initialisation never completes and the `Executor` is never
started.